Fixed quote search to make sort and order persist with search term shifted multiple types to main quote dashboard revised getQuotes to account for all requests rather than having specific pos request in search component
