+++ client/components/quote-sort.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { SortStateUpdate } from './quote-search';
+import { SortStateUpdate } from '../pages/quote-dashboard';
 
 type SortProps = {
   sortType: string;
@@ -24,21 +24,21 @@ export default function QuoteSort (props: SortProps){
   return (
     <>
     <div className="d-flex justify-content-around">
-        <i className="p-1 fa-solid fa-xl fa-filter" onClick={e => toggleModal(!modalShown)}></i>
+        <i className="p-1 fa-solid fa-xl fa-filter" onClick={() => toggleModal(!modalShown)}></i>
         <i className={`p-1 fa-solid fa-xl fa-arrow-${isReversed ? "up" : "down"}`} onClick={reverseList}></i>
     </div>
     <div className={`position-fixed fill-area z-index-10 d-flex justify-content-center align-items-center bg-shadow ${modalShown ? "" : "d-none"}`}>
         <div className="card shadow-sm p-3">
           <label className="d-block">
-            <input className="mx-1" type="radio" name="sortType" value="date" onChange={handleSortSelected} checked={sortType === "date"} />
+            <input className="mx-1" type="radio" name="sortType" value="date" onFocus={handleSortSelected} onChange={handleSortSelected} checked={sortType === "date"} />
             Date Added
           </label>
           <label className="d-block">
-            <input className="mx-1" type="radio" name="sortType" value="page" onChange={handleSortSelected} checked={sortType === "page"} />
+            <input className="mx-1" type="radio" name="sortType" value="page" onFocus={handleSortSelected} onChange={handleSortSelected} checked={sortType === "page"} />
             Page Number
           </label>
           <label className="d-block">
-            <input className="mx-1" type="radio" name="sortType" value="quote" onChange={handleSortSelected} checked={sortType === "quote"} />
+            <input className="mx-1" type="radio" name="sortType" value="quote" onFocus={handleSortSelected} onChange={handleSortSelected} checked={sortType === "quote"} />
             Quote Length
           </label>
         </div>
